# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: X Produçoes e Eventos - API Backend
  version: 1.0.0
  description: >
    API REST para sistema de locação de equipamentos audiovisuais, eventos e gestão de usuários.
    
    Documentação gerada automaticamente.
servers:
  - url: http://localhost:3000/api
    description: Local
  - url: https://api.xproducoes.com.br
    description: Produção
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        avatarUrl:
          type: string
        createdAt:
          type: string
          format: date-time
...
    Booking:
      type: object
      properties:
        id:
          type: string
        eventTitle:
          type: string
        eventDate:
          type: string
          format: date-time
        totalPrice:
          type: number
        status:
          type: string
        clientId:
          type: string
        createdAt:
          type: string
          format: date-time
...
    Contact:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
security:
  - BearerAuth: []
paths:
  /users/register:
    post:
      summary: Registro de novo usuário
      tags: [User]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: Usuário criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/login:
    post:
      summary: Login de usuário
      tags: [User]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login bem-sucedido
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
  /users/profile:
    get:
      summary: Perfil do usuário autenticado
      tags: [User]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Dados do usuário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      summary: Atualizar perfil do usuário
      tags: [User]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                avatar:
                  type: string
                  format: binary
      responses:
        '200':
          description: Perfil atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users:
    get:
      summary: Listar todos os usuários
      tags: [User]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de usuários
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
# Equipment
...
  /bookings:
    get:
      summary: Listar todas as reservas
      tags: [Booking]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de reservas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
    post:
      summary: Criar nova reserva
      tags: [Booking]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Booking'
      responses:
        '201':
          description: Reserva criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
...
  /contacts:
    get:
      summary: Listar mensagens de contato
      tags: [Contact]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de mensagens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contact'
    post:
      summary: Enviar mensagem de contato
      tags: [Contact]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Contact'
      responses:
        '201':
          description: Mensagem enviada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
  /contacts/{id}/read:
    patch:
      summary: Marcar mensagem como lida
      tags: [Contact]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Mensagem marcada como lida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
    delete:
      summary: Remover mensagem de contato
      tags: [Contact]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Mensagem removida

# Notification
...